requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
django-filter==23.3
reportlab==4.0.4
openpyxl==3.1.2